# api/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Asset Wallet Platform API
  version: "1.0"
  description: >
    Internal API for the digital asset wallet platform. Every mutating route requires
    bearer authentication and an Idempotency-Key header (AD-14). Errors use RFC 9457
    application/problem+json.
servers:
  - url: http://localhost:8080/v1
    description: Local dev (docker compose / make run)
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /transfers:
    post:
      operationId: createTransfer
      summary: Move balance from one customer to another, ledger-only, no chain interaction
      description: >
        A single balanced journal entry (debit source, credit destination) commits
        atomically within one (chain, asset) pair — the same granularity as the
        provisioned per-customer accounts (FR4, AD-3, AD-4). No on-chain transaction is
        ever created.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "201":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, invalid chain/asset,
            non-positive amount, or source equals destination
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Source or destination customer has no account for the requested (chain, asset)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            Idempotency-Key reused with a different request body, or a concurrent
            request already used this key to write a journal entry
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: Source account balance is less than the requested amount
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /withdrawals/fee-estimate:
    get:
      operationId: getWithdrawalFeeEstimate
      summary: Query a withdrawal's estimated fee before submitting one
      description: >
        Combines the L2 execution fee and the amortized L1 data-posting fee for the given
        chain (Arbitrum via NodeInterface.gasEstimateComponents(), Base via the
        GasPriceOracle predeploy) — a naive single-number gas estimate would systematically
        undercharge by omitting the L1 component entirely, so both components are always
        returned explicitly alongside their sum, never collapsed into one undifferentiated
        number. No withdrawal resource exists until Story 3.2 — this is a pure,
        unpersisted, read-only computation; nothing is written.
      parameters:
        - name: chain
          in: query
          required: true
          schema:
            type: string
            enum: [base, arbitrum]
        - name: asset
          in: query
          required: true
          schema:
            type: string
            enum: [eth, usdc]
        - name: amount
          in: query
          required: true
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — matching Balance's convention. Must be a positive integer no larger
            than a uint256. Validated but not itself a factor in the returned fee: gas
            cost for a plain transfer or token transfer call does not vary with the
            transferred amount, so the same fee is returned for any valid amount of a
            given chain/asset pair.
          schema:
            type: string
      responses:
        "200":
          description: The withdrawal's estimated fee, split into its L2 and L1 components
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeEstimate"
        "400":
          description: >
            Invalid chain or asset enum value, or amount is missing, non-positive, or not
            a base-10 integer
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: >
            Unexpected server error — e.g. the requested chain/asset combination has no
            token_registry entry for USDC (a misconfigured deployment, never a guessed
            contract address)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /withdrawals:
    post:
      operationId: createWithdrawal
      summary: Request a withdrawal, placing an immediate ledger-only hold on the requested amount
      description: >
        Reserves the requested amount the instant it is accepted (Story 3.2): a single
        balanced journal entry debits the customer's available account and credits its
        hold account for the same (chain, asset) pair — a plain ledger reclassification,
        exactly like an internal transfer, except both accounts belong to the same
        customer. No money leaves the customer and no on-chain transaction is created by
        this endpoint; signing and broadcasting are later stories (3.3-3.5). The
        destination address is validated for structural well-formedness only — never
        checked against a denylist or any other pre-signing policy, which is Story 3.3's
        job.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal accepted; its hold has been placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, invalid chain/asset,
            non-positive amount, or a destination address that is not a well-formed
            0x-prefixed 40-hex-character address
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            Idempotency-Key reused with a different request body, or a concurrent
            request already used this key to write a journal entry
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: >
            Available account balance is less than the requested amount, or (Story 3.3)
            covers the amount but not the amount plus the estimated withdrawal fee
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      summary: Approve an awaiting-approval withdrawal (Story 3.3, FR17, NFR11)
      description: >
        Transitions a withdrawal from "awaiting-approval" to "approved", logging the
        approving operator's actor/timestamp/reason directly on the withdrawal row
        (NFR11). This system has no separate operator-identity/auth tier, so actor is
        caller-supplied in the request body, never derived from the request's auth
        context (a pre-existing, already-logged limitation of the shared-bearer-token
        model). Rejecting an already-approved or non-existent withdrawal is loud
        (404/409), never a silent no-op.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveWithdrawalRequest"
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, or missing/empty
            actor or reason
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No withdrawal with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            The withdrawal is not currently awaiting approval (already approved, still
            created, or a concurrent request already won the approval race), or
            Idempotency-Key reused with a different request body
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    TransferRequest:
      type: object
      required:
        - sourceCustomerId
        - destinationCustomerId
        - chain
        - asset
        - amount
      properties:
        sourceCustomerId:
          type: string
          format: uuid
        destinationCustomerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — matching Balance's convention. Must be a
            positive integer.
    Transfer:
      type: object
      required:
        - id
        - sourceCustomerId
        - destinationCustomerId
        - chain
        - asset
        - amount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        sourceCustomerId:
          type: string
          format: uuid
        destinationCustomerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
        createdAt:
          type: string
          format: date-time
    WithdrawalRequest:
      type: object
      required:
        - customerId
        - chain
        - asset
        - amount
        - destinationAddress
      properties:
        customerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — matching Balance's convention. Must be a
            positive integer.
        destinationAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          minLength: 42
          maxLength: 42
          description: >
            The on-chain address the withdrawal will eventually be sent to. Validated
            for structural well-formedness only (Story 3.2) — never checked against a
            denylist or any other pre-signing policy, which is Story 3.3's job.
    Withdrawal:
      type: object
      required:
        - id
        - customerId
        - chain
        - asset
        - amount
        - destinationAddress
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
        destinationAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          minLength: 42
          maxLength: 42
        status:
          type: string
          enum: [created, awaiting-approval, approved]
          description: >
            "created" is a withdrawal's initial status, immediately superseded in the
            same request by Story 3.3's policy-check-and-route step — no withdrawal is
            ever observable at rest in this status via this API. "awaiting-approval" is
            reached when the requested amount exceeds its (chain, asset)'s configured
            approval threshold (FR17); "approved" is reached either automatically
            (amount at or below threshold) or via POST /withdrawals/{id}/approve. Later
            stories (3.4-3.5) add further status values as they add the transitions that
            produce them.
        approvedAt:
          type: string
          format: date-time
          description: >
            When this withdrawal was approved (Story 3.3, NFR11) — absent for a
            withdrawal still awaiting approval.
        approvedBy:
          type: string
          description: >
            The operator (or, for an auto-approved withdrawal, absent) who approved this
            withdrawal (Story 3.3, NFR11) — caller-supplied at approval time, since this
            system has no separate operator-identity/auth tier.
        approvalReason:
          type: string
          description: >
            The reason given for approving this withdrawal (Story 3.3, NFR11) — absent
            for an auto-approved withdrawal or one still awaiting approval.
        createdAt:
          type: string
          format: date-time
    ApproveWithdrawalRequest:
      type: object
      required:
        - actor
        - reason
      properties:
        actor:
          type: string
          description: >
            The operator approving this withdrawal (NFR11) — required and non-empty;
            this system has no separate operator-identity/auth tier to derive this from.
        reason:
          type: string
          description: The reason for approving this withdrawal (NFR11) — required and non-empty.
...
    FeeEstimate:
      type: object
      required:
        - l2Fee
        - l1Fee
        - totalFee
      properties:
        l2Fee:
          type: string
          description: >
            The L2 execution fee component, in integer base units (wei), encoded as a
            string — never a JSON number — matching Balance's convention.
        l1Fee:
          type: string
          description: >
            The amortized L1 data-posting fee component, in integer base units (wei),
            encoded as a string — never a JSON number.
        totalFee:
          type: string
          description: >
            l2Fee + l1Fee, in integer base units (wei), encoded as a string — never a JSON
            number, and never a naive single-number estimate collapsing the two
            components.
    # RFC 9457 (application/problem+json). Every error response from this service
    # uses this shape, starting this story — see internal/adapter/api/problem.go.
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string